toyota-vehicle-finder/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.9.2
python-multipart==0.0.6
python-dotenv==1.0.0
fastapi-cors==0.0.6
